crawler/requirements.txt
pandas==0.25.3
pymongo==3.10.0
Scrapy==1.8.0
django-environ==0.4.5
requests==2.22.0